+++ tests/fixtures/cassettes/TestGetSinglePhoto.test_is_photo_deleted[16062734376-True].yml
@@ -24,13 +24,13 @@ interactions:
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 15:17:30 GMT
+      - Mon, 29 Apr 2024 15:52:19 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 d4a1f30a438c41cdd21510ef3201c44e.cloudfront.net (CloudFront)
+      - 1.1 7441f523d9aa7a75eb213f3a670e46ac.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - ejYV-dcdkOf06tXHjl7pwyzicncSmbLBxzEFS01QQMc3jfZXyQ_8SA==
+      - eie0BvTRdNu-0segnZmChSxeFiLQAwss_FXgleyQr9TmgPWf_8VL0g==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -41,17 +41,17 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
-        expires=Wed, 29-May-2024 15:17:30 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 15:52:19 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 15:17:30 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 15:52:19 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fba0a-20c66e2c61cf35b93b109c9c;Root=1-662fba0a-6d46734c6f4829dc1e93da97
+      - Self=1-662fc233-69b84f6637ca788f3b37ce5e;Root=1-662fc233-2569a95f424cf3630f6f4c44
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.44.190
+      - serverless-proxy-10.78.27.207
     http_version: HTTP/1.1
     status_code: 200
 version: 1
