+++ tests/fixtures/cassettes/TestGetSinglePhoto.test_is_photo_deleted[53509656752-False].yml
@@ -57,13 +57,13 @@ interactions:
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 15:17:31 GMT
+      - Mon, 29 Apr 2024 15:52:20 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 f60de1480a12e102280c50972fa2a8e8.cloudfront.net (CloudFront)
+      - 1.1 813272f434255b631960acc4a58151c6.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - x8IU5SQc-YRBO-cusjtZfRCAtsbLDFA8-f_Wzhl4o0mPILCEKEO1ag==
+      - fZ4YIhXUbr6de0S-_4VOTjXxFngKBCwx7D25BTFj4EXaP59ynmS_VQ==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -74,17 +74,17 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
-        expires=Wed, 29-May-2024 15:17:31 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 15:52:19 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 15:17:31 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 15:52:19 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fba0b-6ab1eab666809400680afab6;Root=1-662fba0b-1312cca51581046e30545d07
+      - Self=1-662fc233-5ec135280af504b95686c774;Root=1-662fc233-12c18da320835a4d059fe59e
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.10.17
+      - serverless-proxy-10.78.27.207
     http_version: HTTP/1.1
     status_code: 200
 version: 1
