+++ tests/fixtures/cassettes/TestLookupUser.test_lookup_user_by_id.yml
@@ -16,25 +16,21 @@ interactions:
     method: GET
     uri: https://api.flickr.com/services/rest/?method=flickr.people.getInfo&user_id=199258389%40N04
   response:
-    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"ok\">\n<person
-      id=\"199258389@N04\" nsid=\"199258389@N04\" ispro=\"0\" is_deleted=\"0\" iconserver=\"65535\"
-      iconfarm=\"66\" path_alias=\"alexwlchan\" has_stats=\"0\" has_adfree=\"0\" has_free_standard_shipping=\"0\"
-      has_free_educational_resources=\"0\">\n\t<username>alexwlchan</username>\n\t<realname>Alex
-      Chan</realname>\n\t<location />\n\t<description>Tech lead at the Flickr Foundation.</description>\n\t<photosurl>https://www.flickr.com/photos/alexwlchan/</photosurl>\n\t<profileurl>https://www.flickr.com/people/alexwlchan/</profileurl>\n\t<mobileurl>https://www.flickr.com/photos/alexwlchan/</mobileurl>\n\t<photos>\n\t\t<firstdatetaken>2024-02-07
-      06:26:16</firstdatetaken>\n\t\t<firstdate>1707315985</firstdate>\n\t\t<count>1</count>\n\t</photos>\n</person>\n</rsp>\n"
+    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"fail\">\n\t<err
+      code=\"100\" msg=\"Invalid API Key (Key has invalid format)\" />\n</rsp>\n"
     headers:
       Connection:
       - keep-alive
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Tue, 26 Mar 2024 10:48:18 GMT
+      - Mon, 29 Apr 2024 15:48:38 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 718d744faad6ff02c7a7ca517a01865a.cloudfront.net (CloudFront)
+      - 1.1 d4a1f30a438c41cdd21510ef3201c44e.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - XGtPCZ6EsWe_baJ7aT5Cadcrrkcl1ZMJTy9HEggClT69dcgQFU4PDw==
+      - FUm6kGLLRzX-pWZIsjl9pj7oVH7vf9QKBdyJdZbnL7cu2e-5ZKVHeQ==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -43,14 +39,19 @@ interactions:
       - gzip
       server:
       - openresty
+      set-cookie:
+      - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
+        expires=Wed, 29-May-2024 15:48:38 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+      - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
+        expires=Wed, 29-May-2024 15:48:38 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-6602a7f2-55c14ca1144a0fcc2dfc0594;Root=1-6602a7f2-1bc9bbb93731ea5c6d9149ab
+      - Self=1-662fc156-72374ff0019f7ca2344b7d1b;Root=1-662fc156-0d353218441ceae026597d93
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.19.152
+      - serverless-proxy-10.78.11.164
     http_version: HTTP/1.1
     status_code: 200
 version: 1
