+++ tests/fixtures/cassettes/TestLookupUser.test_lookup_user_by_url.yml
@@ -14,77 +14,23 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2F199246608%40N02
-  response:
-    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"ok\">\n<user
-      id=\"199246608@N02\">\n\t<username>cefarrjf87</username>\n</user>\n</rsp>\n"
-    headers:
-      Connection:
-      - keep-alive
-      Content-Type:
-      - text/xml; charset=utf-8
-      Date:
-      - Tue, 26 Mar 2024 10:55:37 GMT
-      Transfer-Encoding:
-      - chunked
-      Via:
-      - 1.1 063bfb014e66ef670fc62ff044660cf2.cloudfront.net (CloudFront)
-      X-Amz-Cf-Id:
-      - NFFntvouzlh2ToWM0XqFPugdELuRM06k1MlXQz2cLv6AqI_NwhljXw==
-      X-Amz-Cf-Pop:
-      - LHR5-P1
-      X-Cache:
-      - Miss from cloudfront
-      content-encoding:
-      - gzip
-      server:
-      - openresty
-      vary:
-      - Accept-Encoding
-      x-flickr-api-request:
-      - Self=1-6602a9a9-5645461874937abf0d3acb48;Root=1-6602a9a9-4498f4804d186612154f62fe
-      x-robots-tag:
-      - noindex
-      x-server:
-      - serverless-proxy-10.78.10.17
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      host:
-      - api.flickr.com
-      user-agent:
-      - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
-        hello@flickr.org)
-    method: GET
     uri: https://api.flickr.com/services/rest/?method=flickr.people.getInfo&user_id=199246608%40N02
   response:
-    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"ok\">\n<person
-      id=\"199246608@N02\" nsid=\"199246608@N02\" ispro=\"0\" is_deleted=\"0\" iconserver=\"0\"
-      iconfarm=\"0\" path_alias=\"\" has_stats=\"0\" has_adfree=\"0\" has_free_standard_shipping=\"0\"
-      has_free_educational_resources=\"0\">\n\t<username>cefarrjf87</username>\n\t<realname>Alex
-      Chan</realname>\n\t<location />\n\t<description />\n\t<photosurl>https://www.flickr.com/photos/199246608@N02/</photosurl>\n\t<profileurl>https://www.flickr.com/people/199246608@N02/</profileurl>\n\t<mobileurl>https://www.flickr.com/photos/199246608@N02/</mobileurl>\n\t<photos>\n\t\t<firstdatetaken>2014-09-18
-      14:11:54</firstdatetaken>\n\t\t<firstdate>1696938538</firstdate>\n\t\t<count>38</count>\n\t</photos>\n</person>\n</rsp>\n"
+    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"fail\">\n\t<err
+      code=\"100\" msg=\"Invalid API Key (Key has invalid format)\" />\n</rsp>\n"
     headers:
       Connection:
       - keep-alive
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Tue, 26 Mar 2024 10:55:37 GMT
+      - Mon, 29 Apr 2024 15:48:37 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 063bfb014e66ef670fc62ff044660cf2.cloudfront.net (CloudFront)
+      - 1.1 281c9390ff02ec3c7b07e745cb742ca6.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - dXYe5ThTBH_BGhOMNLgd_Bq4rl6e3dbJ25uvrEYrT7QigVTDR90Vmw==
+      - 8O3ROMA6piOvhgW3NLYVEr8K1ftznQkXl0TCvobXJOR7qLUWu9A8Tw==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -93,14 +39,19 @@ interactions:
       - gzip
       server:
       - openresty
+      set-cookie:
+      - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
+        expires=Wed, 29-May-2024 15:48:37 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+      - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
+        expires=Wed, 29-May-2024 15:48:37 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-6602a9a9-62e8f8c408d98ec915ad1cf4;Root=1-6602a9a9-132fe20f3a6883730f63e9ea
+      - Self=1-662fc155-6ba920146c68169c773dabe4;Root=1-662fc155-28e3305b04cd924d5a1887a8
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.16.171
+      - serverless-proxy-10.78.32.246
     http_version: HTTP/1.1
     status_code: 200
 version: 1
