+++ tests/fixtures/cassettes/test_get_licenses.yml
