+++ tests/fixtures/cassettes/test_lookup_user_gets_description[user_without_description].yml
@@ -1,0 +1,58 @@
+interactions:
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      connection:
+      - keep-alive
+      host:
+      - api.flickr.com
+      user-agent:
+      - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
+        hello@flickr.org)
+    method: GET
+    uri: https://api.flickr.com/services/rest/?method=flickr.people.getInfo&user_id=46143783%40N04
+  response:
+    content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"ok\">\n<person
+      id=\"46143783@N04\" nsid=\"46143783@N04\" ispro=\"1\" is_deleted=\"0\" iconserver=\"3804\"
+      iconfarm=\"4\" path_alias=\"\" has_stats=\"0\" pro_badge=\"standard\" expire=\"0\"
+      has_adfree=\"0\" has_free_standard_shipping=\"0\" has_free_educational_resources=\"0\">\n\t<username>Barra1man
+      (Busy in the Garden)</username>\n\t<realname>Barra1man (Busy in the Garden)</realname>\n\t<mbox_sha1sum>883bd71722f22dfb348ffae0fd0e1e1a918bc3e7</mbox_sha1sum>\n\t<location
+      />\n\t<timezone label=\"Pacific Time (US &amp; Canada); Tijuana\" offset=\"-08:00\"
+      timezone_id=\"PST8PDT\" timezone=\"5\" />\n\t<description />\n\t<photosurl>https://www.flickr.com/photos/46143783@N04/</photosurl>\n\t<profileurl>https://www.flickr.com/people/46143783@N04/</profileurl>\n\t<mobileurl>https://www.flickr.com/photos/46143783@N04/</mobileurl>\n\t<photos>\n\t\t<firstdatetaken>2010-01-02
+      14:23:09</firstdatetaken>\n\t\t<firstdate>1262470989</firstdate>\n\t\t<count>1440</count>\n\t</photos>\n</person>\n</rsp>\n"
+    headers:
+      Connection:
+      - keep-alive
+      Content-Type:
+      - text/xml; charset=utf-8
+      Date:
+      - Tue, 26 Mar 2024 10:57:02 GMT
+      Transfer-Encoding:
+      - chunked
+      Via:
+      - 1.1 ac0dbffb7a8577e06e873c0fe3eaab42.cloudfront.net (CloudFront)
+      X-Amz-Cf-Id:
+      - _gZiufiEd3B63htY8UHNlnKfntT8P4j5rhgavspdzXuTIm-apj5jgg==
+      X-Amz-Cf-Pop:
+      - LHR5-P1
+      X-Cache:
+      - Miss from cloudfront
+      content-encoding:
+      - gzip
+      server:
+      - openresty
+      vary:
+      - Accept-Encoding
+      x-flickr-api-request:
+      - Self=1-6602a9fe-696d8dd8619d40fc2d961541;Root=1-6602a9fe-3772286a572c9ca805dd8243
+      x-robots-tag:
+      - noindex
+      x-server:
+      - serverless-proxy-10.78.39.10
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
