+++ tests/fixtures/cassettes/test_throws_license_not_found_for_bad_id.yml
